+++ addon/locales/en/forms/fd-setting.js
@@ -85,5 +85,7 @@ export default {
   },
   'custom-message': {
     'validate-ok-header': 'No correction required',
+    'validate-empty-header': 'Project is empty',
+    'validate-empty-message': 'Add classes via «Application model» or «Diagrams»',
   }
 };