+++ addon/locales/ru/components/fd-view-definition-item.js
@@ -1,4 +1,7 @@
 export default {
   'name-placeholder': 'Название атрибута',
   'caption-placeholder': 'Заголовок атрибута',
+  'attribute-show': 'Показать',
+  'attribute-hide': 'Скрыть',
+  'attribute-delete': 'Удалить',
 };