+++ addon/objects/uml-primitives/fd-uml-primitive.js
@@ -233,6 +233,7 @@ joint.highlighters.strokeAndButtons = {
 
     cellView.update();
 
+    cellView.model.set('highlightVel', highlightVel);
     cellView.model.set('highlighted', true);
   },
 
@@ -242,6 +243,7 @@ joint.highlighters.strokeAndButtons = {
     let id = stroke.getHighlighterId(magnetEl, opt);
     this.removeButtons(id);
 
+    cellView.model.set('highlightVel', null);
     cellView.model.set('highlighted', false);
   },