+++ addon/routes/fd-architecture.js
@@ -1,8 +1,9 @@
 import Route from '@ember/routing/route';
+import FdShareLoadData from '../mixins/fd-share-load-data';
 import { inject as service } from '@ember/service';
 import { isBlank } from '@ember/utils';
 
-export default Route.extend({
+export default Route.extend(FdShareLoadData, {
 
   /**
    Service that get current project contexts.