+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
-### Changed
+### Fixed
 * Fixed "Element is no longer attached to DOM" error
 
 ## [1.1.2-beta.6] - 2025-02-04