+++ CHANGELOG.md
@@ -1,8 +1,19 @@
 # Ember Flexberry Designer Changelog
 All notable changes to this project will be documented in this file.
 
+
+
+## [Unreleased]
 ### Added
+* Added the stage ID to queryParams when navigating between sitemap items.
+* Added multiple inheritance from externalinterface
+* Added empty project check for 'Create projections, forms and application' and 'UML validator'
 * Add order in attribute. 
+
+### Changed
+* Added message for view delete
+* Added titles for buttons in view.
+
 
 ## [1.1.2-beta.6] - 2025-02-04
 ### Changed