+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Ember Flexberry Designer Changelog
 All notable changes to this project will be documented in this file.
 
+## [Unrealesed] - 2025-02-04
+### Changed 
+* Changed settng button styles
+* Changed ember-flexberry version to 3.13.2-beta.1
+
 ## [1.1.2-beta.6] - 2025-02-04
 ### Changed
 * Update ember-flexberry to 3.13.1.