+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+* Add multiple view.
 ### Fixed
 * Fixed "Element is no longer attached to DOM" error
 
@@ -9,7 +11,7 @@ All notable changes to this project will be documented in this file.
 * Added the stage ID to queryParams when navigating between sitemap items.
 * Added multiple inheritance from externalinterface
 * Added empty project check for 'Create projections, forms and application' and 'UML validator'
-* Add order in attribute. 
+* Add order in attribute.
 
 ### Changed
 * Added message for view delete