+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
-### Added
-* Add multiple view.
+
 ### Fixed
 * Fixed "Element is no longer attached to DOM" error
 
@@ -12,6 +11,7 @@ All notable changes to this project will be documented in this file.
 * Added multiple inheritance from externalinterface
 * Added empty project check for 'Create projections, forms and application' and 'UML validator'
 * Add order in attribute.
+* Add multiple view.
 
 ### Changed
 * Added message for view delete